+++ product_code/course_service/CHANGELOG.md
@@ -2,7 +2,9 @@
 ## Feature
  - Added Gzipping in Options Header
 ## Refactor
+ - Added NotFound to validationErrors
 ## Bugfix
+ - Fixed empty lecturerId causing an internal 404
 
 # [v.0.8.1](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.7.1...course-v0.8.1) (2020-09-15)
 ## Feature